requirements.txt
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
nuscenes-devkit>=1.1.11
pyquaternion>=0.9.9
